Documents/NetBeansProjects/Tpv/src/test/java/LineaTicketTest.java
/*
 * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
 * Click nbfs://nbhost/SystemFileSystem/Templates/UnitTests/JUnit4TestClass.java to edit this template
 */

import com.mycompany.tpv.LineaTicket;
import org.junit.Test;
import static org.junit.Assert.*;

/**
 *
 * @author alber
 */
public class LineaTicketTest {
    
    public LineaTicketTest() {
    }

    @org.junit.Test
    public void testGetNombreProducto() {
        System.out.println("getNombreProducto");
        LineaTicket instance = null;
        String expResult = "";
        String result = instance.getNombreProducto();
        assertEquals(expResult, result);
        fail("The test case is a prototype.");
    }

    @org.junit.Test
    public void testSetNombreProducto() {
        System.out.println("setNombreProducto");
        String nombreProducto = "";
        LineaTicket instance = null;
        instance.setNombreProducto(nombreProducto);
        fail("The test case is a prototype.");
    }

    @org.junit.Test
    public void testGetPrecioProducto() {
        System.out.println("getPrecioProducto");
        LineaTicket instance = null;
        double expResult = 0.0;
        double result = instance.getPrecioProducto();
        assertEquals(expResult, result, 0);
        fail("The test case is a prototype.");
    }

    @org.junit.Test
    public void testSetPrecioProducto() {
        System.out.println("setPrecioProducto");
        double precioProducto = 0.0;
        LineaTicket instance = null;
        instance.setPrecioProducto(precioProducto);
        fail("The test case is a prototype.");
    }

    @org.junit.Test
    public void testGetCantidadProducto() {
        System.out.println("getCantidadProducto");
        LineaTicket instance = null;
        int expResult = 0;
        int result = instance.getCantidadProducto();
        assertEquals(expResult, result);
        fail("The test case is a prototype.");
    }

    @org.junit.Test
    public void testSetCantidadProducto() {
        System.out.println("setCantidadProducto");
        int cantidadProducto = 0;
        LineaTicket instance = null;
        instance.setCantidadProducto(cantidadProducto);
        fail("The test case is a prototype.");
    }
    
    @Test
    public void testConstructor() {
        LineaTicket linea = new LineaTicket("Producto1", 10.5, 3);
        assertEquals("Producto1", linea.getNombreProducto());
        assertEquals(10.5, linea.getPrecioProducto(), 0.001);
        assertEquals(3, linea.getCantidadProducto());
    }
    
    @Test
    public void testGettersAndSetters() {
        LineaTicket linea = new LineaTicket("Producto2", 20.75, 2);
        linea.setNombreProducto("NuevoProducto");
        linea.setPrecioProducto(15.0);
        linea.setCantidadProducto(5);
        
        assertEquals("NuevoProducto", linea.getNombreProducto());
        assertEquals(15.0, linea.getPrecioProducto(), 0.001);
        assertEquals(5, linea.getCantidadProducto());
    }
    
    @Test
    public void testFuncionalidadGeneral() {
        LineaTicket linea = new LineaTicket("Producto3", 30.0, 1);
        assertEquals("Producto3", linea.getNombreProducto());
        assertEquals(30.0, linea.getPrecioProducto(), 0.001);
        assertEquals(1, linea.getCantidadProducto());
    }
    
    @Test
    public void testValoresLimite() {
        LineaTicket linea = new LineaTicket("Producto4", 40.0, 0);
        assertEquals("Producto4", linea.getNombreProducto());
        assertEquals(40.0, linea.getPrecioProducto(), 0.001);
        assertEquals(0, linea.getCantidadProducto());
    }
    
}
